Add support for a custom data dir for Linux system installs
An optional KRONIAX_DATADIR parameter can be passed to CMake to initialise
a custom data dir (e.g. /usr/share/games/kroniax for system-wide installs
on Linux). If the optional parameter is not passed, the data dir defaults
to "./data" as previously.

To achieve this, a Helper::getData function has been added in new global.hpp
header, which can be accessed from any source file that requires access to
data files.

Fixes #16.